Changed the model of the wind speed. Now picks a random wind speed from an average wind speed distribution each time the current wind speed is calculated
